[librptexture] ImageDecoder::fromGcnCI4(): Change the size parameters from `int` to `size_t`.
This matches the other ImageDecoder functions.

This was changed in the other the ImageDecoder functions in
commit fc206f30de9d8f010a91d8c7db3472c72960d6d7 (2022). The CI4 function
was added in 2023, but for some reason I used `int` instead of `size_t`.

Commit fc206f30de9d8f010a91d8c7db3472c72960d6d7:
([librptexture] ImageDecoder: Change `img_siz` and `pal_siz` from `int` to `size_t`.)

Commit 9283ec808e113c124f5d1fc9c07aaa622eac1550:
([librptexture] SegaPVR: Initial implementation of GameCube CI4.)

+++ src/librptexture/decoder/ImageDecoder_GCN.cpp
@@ -284,18 +284,18 @@ rp_image_ptr fromGcnI8(int width, int height,
  * @return rp_image, or nullptr on error.
  */
 rp_image_ptr fromGcnCI4(int width, int height,
-	const uint8_t *RESTRICT img_buf, int img_siz,
-	const uint16_t *RESTRICT pal_buf, int pal_siz)
+	const uint8_t *RESTRICT img_buf, size_t img_siz,
+	const uint16_t *RESTRICT pal_buf, size_t pal_siz)
 {
 	// Verify parameters.
 	assert(img_buf != nullptr);
 	assert(pal_buf != nullptr);
 	assert(width > 0);
 	assert(height > 0);
-	assert(img_siz >= ((width * height) / 2));
+	assert(img_siz >= (((size_t)width * (size_t)height) / 2));
 	assert(pal_siz >= 16*2);
 	if (!img_buf || !pal_buf || width <= 0 || height <= 0 ||
-	    img_siz < ((width * height) / 2) || pal_siz < 16*2)
+	    img_siz < (((size_t)width * (size_t)height) / 2) || pal_siz < 16*2)
 	{
 		return nullptr;
 	}
